Fix typehinting bugs
Also marked confusing mypy linting with comments.

+++ model/metric.py
@@ -5,8 +5,8 @@ from torch import Tensor
 def my_metric(output: Tensor, target: Tensor) -> float:
     with torch.no_grad():
         pred: Tensor = torch.argmax(output, dim=1)
-        assert pred.shape[0] == len(target)
-        correct: int = 0
+        assert pred.shape[0] == len(target)  # mypy error, len(Tensor) always returns # of first dim
+        correct: float = 0
         correct += torch.sum(pred == target).item()
     return correct / len(target)
 
@@ -15,7 +15,7 @@ def my_metric2(output: Tensor, target: Tensor, k: int = 3) -> float:
     with torch.no_grad():
         pred: Tensor = torch.topk(output, k, dim=1)[1]
         assert pred.shape[0] == len(target)
-        correct: int = 0
+        correct: float = 0
 
         i: int
         for i in range(k):
